New bot with enhanced signalling and bug fixes
- New bot that allows rapid development: Allows you to skip setup steps
and just enters the "TEST" state that results in the "test" method being
called from the loop.
- Signalling via "BWB" protocol, namely the use of a white section
between two black sections.  To create such a section of track, make the
black line double-thickness then place a 75% thickness piece of white in
the middle.  Is pretty accurate at 5mm signal sections with 10mm gaps.
Big advantage over white-strip signalling is that it doesn't tend to
interpret corners as signal sections.  Tested at 100 speed.
- Fixed bug whereby after a period of time the robot would seem to enter
a state where it always said the last message was "1".  This was caused
by use of unsigned int for time values whereas millis() returns unsigned
long.  So we were getting overflow after a minute or so of running the
'bot and that was resulting in it never satisfying the MessageReceived()
criteria.  Took me ages to find that, but was obvious as soon as I
checked the Arduino help for the millis() function.