Multi-message State Control
- Added framework for state changes based on multiple levels of
messages.  For example, if a "2" is received, it knows there is a corner
ahead and expects a second message telling it whether the bend will be
to the right or the left [can be used to improve line-following on tight
bends or at high speeds].
- Added ability to select SLOW (button A - speed 50), NORMAL (button B -
speed 100) or FAST (button C  - speed 200) after calibrating.
- Added automatic adjustment of message expiry window based on speed.
So at faster speeds, messages are read faster and vice versa.
- Added pre-printed message strips for turning left (2-1)and right
(2-2), reversing (1-1), setting speed to CAREFUL (1-2) and returning to
normal when in a sub-state (1)