Reversing using proportional and derivative (kind-of)
- followLineInReverse() using proportional and derivative to control
power delta between wheels.  Still not great - we need some kind of
dampener because it oscillates back and forth across the line with ever
increasing amplitude (i.e: it is over-correcting - we need it to lessen
the power differential as it is about to rejoin the line again after
being off it).  This is a difficult challenge!