Tested on see-saw and ramp
- Refined Proportional and Derivative parameters (Proportional / 7,
Derivative * 4)
- This version works well going down a steep ramp
- Also crosses a narrow see-saw with no problems
- Doesn't seem to handle 90-degree turns, though I have seen it do these
with no problem at 255 speed with these parameters, so perhaps it has
something to do with calibration [confirmed signal-checking has no
effect]
- Some refactoring (introduction of stopAndWait() function and removal
of many magic constants)
- Addition to followLine() and followLineSimpleMode() of "drive
straight" rule when on a signal.  This greatly improves line-tracking
when on signal sections.
- Reversing still lets it down.  Tried having a section where it stops
and reverses but it continually failed to receive the "END_REVERSE"
message because it was too far off the line.  However, given excellent
performance on ramps, reversing may no longer be required.